Create RestaurantIndex
Code from my personal website. Index page for the restaurants. This runs a for loop of a list of restaurants and displays bootstrap cards with an image of the restaurant, the name, and type of restaurant (category). If image list is null then default image is displayed.

I may wrap this in a try catch statement